Improves the handling of segmented headers with large content body.
In general, uWebSockets' HTTP handling is quite narrow---it misbehaves
when the headers exceed 4 KiB for example. There may be some arguments
for this kind of behavior but it will probably make uWebSockets
inappropriate for hybrid HTTP-WebSockets tasks.

That said, uWebSockets _does_ try to handle segmented headers (which
are nowadays not really recommended for clients). But there's a catch:
when segmented headers are followed by large content body, TCP is free
to combine the later part of headers with the beginning of the body.
Since this can make the data larger than 4 KiB the server gives up.
This commit fixes this behavior by introducing the "additional" buffer
that is never consulted for headers but considered to be a body once
the headers are complete.

It should be noted that the current support of pipelined request is
barely usable (there is the same kind of problem with buffer scaling);
while related, this commit didn't try to touch that.